% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSampling.R
\name{buildmodeldesign}
\alias{buildmodeldesign}
\title{Stepped wedge design model}
\usage{
buildmodeldesign(I, TP, mu, theta, beta.j, sigma.alpha, X.i.j.0, X.i.j = NULL)
}
\arguments{
\item{I}{number of clusters (design parameter)}

\item{TP}{number of timepoints (design parameter)}

\item{mu}{baseline mean (model parameter)}

\item{theta}{treatment effect (model parameter)}

\item{beta.j}{vector of time trents (model parameter)}

\item{sigma.alpha}{between cluster variability as standard deviation (model parameter)}

\item{X.i.j.0}{assumed treatment model matrix for a SWD study (model parameter)}

\item{X.i.j}{data model matrix of real intervention implementation (model parameter), by default (NULL) the same as the X.i.j.0}
}
\value{
Data frame with mean intensities corresponding to the SWD model and full model parameter information
}
\description{
Performs the mean values of a given Stepped wedge design (SWD)
}
\examples{
noCl<-10
noT<-6
switches<-2
DM<-designMatrix.SWD(noCl,noT,switches)
buildmodeldesign(I=noCl,TP=noT, mu=0,theta=1,beta.j=rep(1,noT),sigma.alpha=0.5, X.i.j.0=DM)
}

