% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerSWD.R
\name{calcPower.Hussey_SWD}
\alias{calcPower.Hussey_SWD}
\title{Power calculation within stepped wedge design model by Hussey et.al}
\usage{
calcPower.Hussey_SWD(ThetaEst, alpha = 0.05, Design, sigmaq, tauq, time)
}
\arguments{
\item{ThetaEst}{expected treatment effect}

\item{alpha}{singificance level (by default 0.05)}

\item{Design}{design matrix for a given SWD model}

\item{sigmaq}{within cluster variance}

\item{tauq}{between cluster variance}

\item{time}{a logical (FALSE, if no time trends are expected, otherwise TRUE)}
}
\value{
Aproximated power of two tailed test, although the design matrix is fractionated, then power is not valid
#@references \cite{Hussey2007}
}
\description{
Calculation of power for a lmm with cluster as random effect, fixed timepoint effects, but set to null, TP number of timepoints, I number of cluster. The design matrix has to be coded by zeros and ones.
}
\examples{
noCl<-10
noT<-6
switches<-2
DM<-designMatrix.SWD(noCl,noT,switches)
sigma.e <- 2
sigma.alpha <- 2    
calcPower.Hussey_SWD(ThetaEst=1,Design=DM, sigmaq=sigma.e^2, tauq=sigma.alpha^2, time=FALSE)
calcPower.Hussey_SWD(ThetaEst=1,Design=v, sigmaq=sigma.e^2, tauq=sigma.alpha^2, time=TRUE)
}

