% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSampling.R
\name{Data.loss.SWD}
\alias{Data.loss.SWD}
\title{For given information of loss sampling response of individuals within a SWD model}
\usage{
Data.loss.SWD(data.all, I, TP, B = "0", C = 0, D = 0)
}
\arguments{
\item{data.all}{Sampled data (response) for given numbers of individuals by given group means within a SWD model}

\item{I}{number of clusters (design parameter)}

\item{TP}{number of timepoints (design parameter)}

\item{B}{timepoint of cluster loss with 4 possibilities: "0": default - no cluster at no timepoint get lost, "1" - Cluster missing at random from timepoint 2 untill TP, "2" - Cluster is missing at beginning (1/3 of timepoints after the first), "3" - Cluster is missing at end (1/3 of the last timepoints).}

\item{C}{number of cluster loss, by default zero. If a cluster get lost from time point i, all indiviual responses of that cluster will be deleted from timepoint i until timpeoint TP (end).}

\item{D}{number of individuals loss, by default zero. If not zero, then individual responses to delete are selected at random from timepoints and clusters.}
}
\value{
Data frame with individuals intensities corresponds to the SWD model and full model parameter information and derivation information
}
\description{
Data for given numbers of individuals by given group means within a SWD model and derivations with loss of data
}
\examples{
noCl<-10
noT<-6
switches<-2
DM<-designMatrix.SWD(noCl,noT,switches)
#cross-sectional SWD (10 cluster and 6 time points)
#no derivation from perfect 100 percent effectiveness pattern
#no data loss (no missing)
data<-SWD.datasampling(I=noCl,TP=noT, mu=0,theta=1,beta.j=rep(1,noT),sigma.alpha=0.5, X.i.j.0=DM,N=10,sigma.e=1)
#no missing in data
Data.loss.SWD(data.all=data, I=noCl,TP=noT)
#missing individuals
Data.loss.SWD(data.all=data, I=noCl,TP=noT, D=5)
#missing 2 cluster at random
Data.loss.SWD(data.all=data, I=noCl,TP=noT, B="1", C=2)
}

