% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatrix.R
\name{designMatrix.SWD}
\alias{designMatrix.SWD}
\title{Design matrix for SWD model}
\usage{
designMatrix.SWD(nC, nT, nSw)
}
\arguments{
\item{nC}{Number of clusters}

\item{nT}{Number of timepoint}

\item{nSw}{number of clusters switches from control to treatment at each timepoint}
}
\value{
design matrix of size noCluster x noTimepoints
}
\description{
Creates a design matrix for stepped wedge design (SWD) given the design parameters
}
\examples{
designMatrix.SWD(5,6,1)
designMatrix.SWD(10,6,2)
}

