% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatrix.R
\name{implemMatrix.SWD}
\alias{implemMatrix.SWD}
\title{Design matrix for SWD model under an intervention effectiveness pattern}
\usage{
implemMatrix.SWD(nC, nT, nSw, pattern)
}
\arguments{
\item{nC}{Number of clusters}

\item{nT}{Number of timepoint}

\item{nSw}{number of clusters switches from control to treatment at each timepoint}

\item{pattern}{a vector for intervention effectiveness pattern gives the derivation from 100 percent effectiveness over time}
}
\value{
intervention effectivness pattern matrix as design matrix of size noCluster x noTimepoints
}
\description{
Creates a implementation matrix for a given stepped wedge design and intervention effectiveness pattern
}
\examples{
implemMatrix.SWD(5,6,1, c(seq(0.4,1,0.2),1))
implemMatrix.SWD(10,6,2, c(seq(1,0.4,-0.2),1))
}

