% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSampling.R
\name{sampleData}
\alias{sampleData}
\title{Sampling Response of individuals within a SWD model}
\usage{
sampleData(I, TP, mu, theta, beta.j, sigma.alpha, X.i.j.0, X.i.j = NULL, N,
  sigma.e, sigma.ind = NULL)
}
\arguments{
\item{I}{number of clusters (design parameter)}

\item{TP}{number of timepoints (design parameter)}

\item{mu}{baseline mean (model parameter)}

\item{theta}{treatment effect (model parameter)}

\item{beta.j}{vector of time trents (model parameter)}

\item{sigma.alpha}{between cluster variability as standard deviation (model parameter)}

\item{X.i.j.0}{assumed treatment model matrix for a SWD study (model parameter)}

\item{X.i.j}{data model matrix of real intervention implementation (model parameter), by default (NULL) the same as the X.i.j.0}

\item{N}{number of individuals (fixed) for all clusters and timepoints}

\item{sigma.e}{random error variability as standard deviation (model parameter)}

\item{sigma.ind}{individual variability as standard deviation (model parameter), if it is an longitudinal model, by default (NULL) it is an cross-sectional model}
}
\value{
Data frame with individuals intensities corresponds to the SWD model and full model parameter information
}
\description{
Sample data (response) for given numbers of individuals by given group means within a SWD model
}
\examples{
noCl<-10
noT<-6
switches<-2
DM<-designMatrix.SWD(noCl,noT,switches)
sampleData(I=noCl,TP=noT, mu=0,theta=1,beta.j=rep(1,noT),sigma.alpha=0.5, X.i.j.0=DM,N=10,sigma.e=1)
}

