% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationSWD.R
\name{simABC}
\alias{simABC}
\title{Simulate data and estimate lmm within stepped wedge design model}
\usage{
simABC(I, TP, mu, theta, beta.j, sigma.alpha, X.i.j.0, N, sigma.e,
  sigma.ind = NULL, A = NULL, B = "0", C = 0, D = 0)
}
\arguments{
\item{I}{number of clusters (design parameter)}

\item{TP}{number of timepoints (design parameter)}

\item{mu}{baseline mean (model parameter)}

\item{theta}{treatment effect (model parameter)}

\item{beta.j}{vector of time trents (model parameter)}

\item{sigma.alpha}{between cluster variability as standard deviation (model parameter)}

\item{X.i.j.0}{assumed treatment model matrix for a SWD study (model parameter)}

\item{N}{number of individuals (fixed) for all clusters and timepoints}

\item{sigma.e}{random error variability as standard deviation (model parameter)}

\item{sigma.ind}{individual variability as standard deviation (model parameter), if it is an longitudinal model, by default (NULL) it is an cross-sectional model}

\item{A}{derivation from perfect 100 percent effectiveness pattern (simulation parameter)}

\item{B}{timepoint of cluster loss (simulation parameter) with 4 possibilities: "0": default - no cluster at no timepoint get lost, "1" - Cluster missing at random from timepoint 2 untill TP, "2" - Cluster is missing at beginning (1/3 of timepoints after the first), "3" - Cluster is missing at end (1/3 of the last timepoints).}

\item{C}{number of cluster loss (simulation parameter), by default zero. If a cluster get lost from time point i, all indiviual responses of that cluster will be deleted from timepoint i until timpeoint TP (end).}

\item{D}{number of individuals loss (simulation parameter), by default zero. If not zero, then individual responses to delete are selected at random from timepoints and clusters.}
}
\value{
treatment and time effects estimated with linear mixed model  
#@examples
}
\description{
For a given one scenario build mean intensities of SWD model, sample data and estimate treatment parameter using a linear mixed model
TP number of timepoints, I number of cluster. The design matrix has to be coded by zeros and ones.
}

