% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSampling.R
\name{SWD.datasampling}
\alias{SWD.datasampling}
\title{Sampling Response of individuals within a SWD model}
\usage{
SWD.datasampling(I, TP, mu, theta, beta.j, sigma.alpha, X.i.j.0, N, sigma.e,
  sigma.ind = NULL, A = NULL, B = "0", C = 0, D = 0)
}
\arguments{
\item{I}{number of clusters (design parameter)}

\item{TP}{number of timepoints (design parameter)}

\item{mu}{baseline mean (model parameter)}

\item{theta}{treatment effect (model parameter)}

\item{beta.j}{vector of time trents (model parameter)}

\item{sigma.alpha}{between cluster variability as standard deviation (model parameter)}

\item{X.i.j.0}{assumed treatment model matrix for a SWD study (model parameter)}

\item{N}{number of individuals (fixed) for all clusters and timepoints}

\item{sigma.e}{random error variability as standard deviation (model parameter)}

\item{sigma.ind}{individual variability as standard deviation (model parameter), if it is an longitudinal model, by default (NULL) it is an cross-sectional model}

\item{A}{derivation from perfect 100 percent effectiveness pattern}

\item{B}{timepoint of cluster loss with 4 possibilities: "0": default - no cluster at no timepoint get lost, "1" - Cluster missing at random from timepoint 2 untill TP, "2" - Cluster is missing at beginning (1/3 of timepoints after the first), "3" - Cluster is missing at end (1/3 of the last timepoints).}

\item{C}{number of cluster loss, by default zero. If a cluster get lost from time point i, all indiviual responses of that cluster will be deleted from timepoint i until timpeoint TP (end).}

\item{D}{number of individuals loss, by default zero. If not zero, then individual responses to delete are selected at random from timepoints and clusters.}

\item{X.i.j}{data model matrix of real intervention implementation (model parameter), by default (NULL) the same as the X.i.j.0}
}
\value{
Data frame with individuals intensities corresponds to the SWD model and full model parameter information and derivation information
}
\description{
Sample data (response) for given numbers of individuals by given group means within a SWD model and derivations
}
\examples{
noCl<-10
noT<-6
switches<-2
DM<-designMatrix.SWD(noCl,noT,switches)
#cross-sectional SWD (10 cluster and 6 time points)
#no derivation from perfect 100 percent effectiveness pattern
#no data loss (no missing)
#SWD.datasampling(I=noCl,TP=noT, mu=0,theta=1,beta.j=rep(1,noT),sigma.alpha=0.5, X.i.j.0=DM,N=10,sigma.e=1)
#cross-sectional SWD (10 cluster and 6 time points)
#no derivation from perfect 100 percent effectiveness pattern
#missing individuals
#SWD.datasampling(I=noCl,TP=noT, mu=0,theta=1,beta.j=rep(1,noT),sigma.alpha=0.5, X.i.j.0=DM,N=10,sigma.e=1, D=5)
#cross-sectional SWD (10 cluster and 6 time points)
#no derivation from perfect 100 percent effectiveness pattern
#missing 2 cluster at random
#SWD.datasampling(I=noCl,TP=noT, mu=0,theta=1,beta.j=rep(1,noT),sigma.alpha=0.5, X.i.j.0=DM,N=10,sigma.e=1 ,B="1", C=2)
#longitudinal SWD (10 cluster and 6 time points)
#no derivation from perfect 100 percent effectiveness pattern
#no data loss (no missing)
#SWD.datasampling(I=noCl,TP=noT, mu=0,theta=1,beta.j=rep(1,noT),sigma.alpha=0.5, X.i.j.0=DM,N=10,sigma.e=1, sigma.ind=0.5)
}

